app/requirements.txt
fastapi
uvicorn[standard]
asyncpg
gunicorn
sqlalchemy
psycopg2-binary  
pydantic
python-jose[cryptography]
py-bcrypt
python-decouple
python-multipart
google-genai
alembic
unidecode